# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest API documentation, generated based on annotations and swag library",
        "title": "Lab3 Rest API",
        "contact": {}
    },
    "host": "localhost:8080",
    "paths": {
        "/authenticate": {
            "post": {
                "description": "authenticate user in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get token",
                "parameters": [
                    {
                        "description": "Submit",
                        "name": "authenticationForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Token"
                        }
                    }
                }
            }
        },
        "/boards": {
            "get": {
                "description": "retrieves all saved board entities from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "get all possible boards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.BoardDto"
                            }
                        }
                    }
                }
            }
        },
        "/boards/{boardId}": {
            "get": {
                "description": "respond board dto by provided id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "get possible board by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "use id",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BoardDto"
                        }
                    }
                }
            }
        },
        "/calculations": {
            "post": {
                "description": "receives a 2D array from request body and search on it a circle loop, depends on default conditions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculations"
                ],
                "summary": "find circle loop",
                "parameters": [
                    {
                        "description": "Submit",
                        "name": "calculations",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BoardDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BoardDto"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "registers new user with provided date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "register as new user",
                "parameters": [
                    {
                        "description": "Submit",
                        "name": "registrationForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserDto"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.BoardDto": {
            "type": "object",
            "properties": {
                "board": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "model.UserDto": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}